Miscellanea

DiffinDiffsBase.cb — Method

cb(args...)

Construct a vector from ConstantTerms provided as arguments. This method is useful for working with @formula.

Note that @formula does not accept function calls with empty argument. Hence, cb() cannot be used to create an empty vector [] in the expression passed to @formula. However, passing cb (without the parenthesis) to @formula will return [] due to unpack.

DiffinDiffsBase.exampledata — Method

exampledata(name::Union{Symbol,String})

Return a CSV.File containing the example dataset with the specified name.

DiffinDiffsBase.exampledata — Method

exampledata()

Return the names of available example datasets.

DiffinDiffsBase.kwarg — Method

kwarg(v)

Return a key-value Pair with the key being a keyword argument name and the value being v. The key is determined by the type of v.

DiffinDiffsBase.unpack — Method

unpack(t::ConstantTerm)

Return the value represented by t.

DiffinDiffsBase.unpack — Method

unpack(t::Term)

If t.sym is the name of an object defined in Main, either call its method with no argument if it exists or return the object itself. Otherwise, return t.sym.

DiffinDiffsBase.unpack — Method

unpack(t::FunctionTerm{F})

Call the function represented by t.

DiffinDiffsBase.@fieldequal — Macro

@fieldequal Supertype

Define a method of == for all subtypes of Supertype such that == returns true if two instances have the same field values.

DiffinDiffsBase.@unpack — Macro

@unpack functionname

Define a method of functionname that accepts terms generated by @formula as arguments. This method can be used to translate terms into arguments that match the other methods of functionname. Each term is processed by unpack and kwarg.